+++ infra/modules/worklytics-connector-specs/main.tf
@@ -157,8 +157,8 @@ locals {
 
   jira_cloud_id                 = coalesce(var.jira_cloud_id, "YOUR_JIRA_CLOUD_ID")
   jira_example_issue_id         = coalesce(var.jira_example_issue_id, var.example_jira_issue_id, "YOUR_JIRA_EXAMPLE_ISSUE_ID")
-  github_api_host               = coalesce(var.github_api_host, "api.github.com")
   github_installation_id        = coalesce(var.github_installation_id, "YOUR_GITHUB_INSTALLATION_ID")
+  github_enterprise_server_host = coalesce(var.github_api_host, var.github_enterprise_server_host, "YOUR_GITHUB_ENTERPRISE_SERVER_HOST")
   github_organization           = coalesce(var.github_organization, "YOUR_GITHUB_ORGANIZATION_NAME")
   github_example_repository     = coalesce(var.github_example_repository, "YOUR_GITHUB_EXAMPLE_REPOSITORY_NAME")
   salesforce_example_account_id = coalesce(var.salesforce_example_account_id, "{ANY ACCOUNT ID}")
@@ -374,7 +374,7 @@ EOT
       display_name : "Github Enterprise"
       identifier_scope_id : "github"
       worklytics_connector_name : "Github Enterprise via Psoxy"
-      target_host : local.github_api_host
+      target_host : "api.github.com"
       source_auth_strategy : "oauth2_refresh_token"
       secured_variables : [
         {
@@ -406,7 +406,7 @@ EOT
       environment_variables : {
         GRANT_TYPE : "certificate_credentials"
         TOKEN_RESPONSE_TYPE : "GITHUB_ACCESS_TOKEN"
-        REFRESH_ENDPOINT : "https://${local.github_api_host}/app/installations/${local.github_installation_id}/access_tokens"
+        REFRESH_ENDPOINT : "https://api.github.com/app/installations/${local.github_installation_id}/access_tokens"
         USE_SHARED_TOKEN : "TRUE"
       }
       settings_to_provide = {
@@ -474,13 +474,13 @@ https://github.com/organizations/{YOUR ORG}/settings/installations/{INSTALLATION
 
 EOT
     }
-    github-non-enterprise = {
-      source_kind : "github",
-      worklytics_connector_id : "github-free-team-psoxy"
-      display_name : "Github"
+    github-enterprise-server = {
+      source_kind : "github-enterprise-server",
+      worklytics_connector_id : "github-enterprise-server-psoxy"
+      display_name : "Github Enterprise Server"
       identifier_scope_id : "github"
-      worklytics_connector_name : "Github Free/Professional via Psoxy"
-      target_host : local.github_api_host
+      worklytics_connector_name : "Github Enterprise Server via Psoxy"
+      target_host : local.github_enterprise_server_host
       source_auth_strategy : "oauth2_refresh_token"
       secured_variables : [
         {
@@ -512,7 +512,114 @@ EOT
       environment_variables : {
         GRANT_TYPE : "certificate_credentials"
         TOKEN_RESPONSE_TYPE : "GITHUB_ACCESS_TOKEN"
-        REFRESH_ENDPOINT : "https://${local.github_api_host}/app/installations/${local.github_installation_id}/access_tokens"
+        REFRESH_ENDPOINT : "https://${local.github_enterprise_server_host}/app/installations/${local.github_installation_id}/access_tokens"
+        USE_SHARED_TOKEN : "TRUE"
+      }
+      settings_to_provide = {
+        "GitHub Organization" = local.github_organization
+      }
+      reserved_concurrent_executions : null
+      example_api_calls_user_to_impersonate : null
+      example_api_calls : [
+        "/api/v3/orgs/${local.github_organization}/repos",
+        "/api/v3/orgs/${local.github_organization}/members",
+        "/api/v3/orgs/${local.github_organization}/teams",
+        "/api/v3/orgs/${local.github_organization}/audit-log",
+        "/api/v3/repos/${local.github_organization}/${local.github_example_repository}/events",
+        "/api/v3/repos/${local.github_organization}/${local.github_example_repository}/commits",
+        "/api/v3/repos/${local.github_organization}/${local.github_example_repository}/issues",
+        "/api/v3/repos/${local.github_organization}/${local.github_example_repository}/pulls",
+      ]
+      external_token_todo : <<EOT
+  1. You have to populate `github_enterprise_server_host` variable in Terraform with the hostname of your Github Enterprise Server (example: `github.your-company.com`).
+  2. From your organization, register a [GitHub App](https://docs.github.com/en/enterprise-server@3.11/apps/creating-github-apps/registering-a-github-app/registering-a-github-app#registering-a-github-app)
+    with following permissions with **Read Only**:
+    - Repository:
+      - Contents: for reading commits and comments
+      - Issues: for listing issues, comments, assignees, etc.
+      - Metadata: for listing repositories and branches
+      - Pull requests: for listing pull requests, reviews, comments and commits
+    - Organization
+      - Administration: for listing events from audit log
+      - Members: for listing teams and their members
+
+  NOTES:
+    - We assume that ALL the repositories are going to be listed **should be owned by the organization, not the users**.
+    - Enterprise Cloud is required for this connector.
+
+  Apart from Github instructions please review the following:
+  - "Homepage URL" can be anything, not required in this flow but required by Github.
+  - Webhooks check can be disabled as this connector is not using them
+  - Keep `Expire user authorization tokens` enabled, as GitHub documentation recommends
+  3. Once is created please generate a new `Private Key`.
+  4. It is required to convert the format of the certificate downloaded from PKCS#1 in previous step to PKCS#8. Please run following command:
+```shell
+openssl pkcs8 -topk8 -inform PEM -outform PEM -in {YOUR DOWNLOADED CERTIFICATE FILE} -out gh_pk_pkcs8.pem -nocrypt
+```
+
+**NOTES**:
+ - If the certificate is not converted to PKCS#8 connector will NOT work. You might see in logs a Java error `Invalid PKCS8 data.` if the format is not correct.
+ - Command proposed has been successfully tested on Ubuntu; it may differ for other operating systems.
+
+  5. Install the application in your organization.
+     Go to your organization settings and then in "Developer Settings". Then, click on "Edit" for your "Github App" and once you are in the app settings, click on "Install App" and click on the "Install" button. Accept the permissions to install it in your whole organization.
+  6. Once installed, the `installationId` is required as it needs to be provided in the proxy as parameter for the connector in your Terraform module. You can go to your organization settings and
+click on `Third Party Access`. Click on `Configure` the application you have installed in previous step and you will find the `installationId` at the URL of the browser:
+```
+https://{YOUR GITHUB HOST}/organizations/{YOUR ORG}/settings/installations/{INSTALLATION_ID}
+```
+  Copy the value of `installationId` and assign it to the `github_installation_id` variable in Terraform. You will need to redeploy the proxy again if that value was not populated before.
+
+**NOTE**:
+ - If `github_installation_id` is not set, authentication URL will not be properly formatted and you will see *401: Unauthorized* when trying to get an access token.
+ - If you see *404: Not found* in logs please review the *IP restriction policies* that your organization might have; that could cause connections from psoxy AWS Lambda/GCP Cloud Functions be rejected.
+
+  6. Update the variables with values obtained in previous step:
+     - `PSOXY_GITHUB_CLIENT_ID` with `App ID` value. **NOTE**: It should be `App Id` value as we are going to use authentication through the App and **not** *client_id*.
+     - `PSOXY_GITHUB_PRIVATE_KEY` with content of the `gh_pk_pkcs8.pem` from previous step. You could open the certificate with VS Code or any other editor and copy all the content *as-is* into this variable.
+  7. Once the certificate has been uploaded, please remove {YOUR DOWNLOADED CERTIFICATE FILE} and `gh_pk_pkcs8.pem` from your computer or store it in a safe place.
+
+EOT
+    }
+    github-non-enterprise = {
+      source_kind : "github",
+      worklytics_connector_id : "github-free-team-psoxy"
+      display_name : "Github"
+      identifier_scope_id : "github"
+      worklytics_connector_name : "Github Free/Professional/Team via Psoxy"
+      target_host : "api.github.com"
+      source_auth_strategy : "oauth2_refresh_token"
+      secured_variables : [
+        {
+          name : "ACCESS_TOKEN"
+          writable : true
+          sensitive : true
+          value_managed_by_tf : false
+        },
+        {
+          name : "PRIVATE_KEY"
+          writable : false
+          sensitive : true
+          value_managed_by_tf : false
+        },
+        {
+          name : "CLIENT_ID"
+          writable : false
+          sensitive : true
+          value_managed_by_tf : false
+        },
+        {
+          name : "OAUTH_REFRESH_TOKEN"
+          writable : true
+          lockable : true
+          sensitive : true
+          value_managed_by_tf : false
+        }
+      ],
+      environment_variables : {
+        GRANT_TYPE : "certificate_credentials"
+        TOKEN_RESPONSE_TYPE : "GITHUB_ACCESS_TOKEN"
+        REFRESH_ENDPOINT : "https://api.github.com/app/installations/${local.github_installation_id}/access_tokens"
         USE_SHARED_TOKEN : "TRUE"
       }
       settings_to_provide = {