+++ infra/modules/worklytics-connector-specs/variables.tf
@@ -88,10 +88,17 @@ variable "jira_example_issue_id" {
   description = "If using Jira Server/Cloud connector, provide id of an issue for only to be used as part of example calls for Jira (ex: ETV-12)"
 }
 
+# DEPRECATED
 variable "github_api_host" {
   type        = string
   default     = null
-  description = "(Only required if using Github connector for on premises) Host of the Github instance (ex: github.mycompany.com). If not provided, it will use 'api.github.com' as default value"
+  description = "DEPRECATED; use `github_enterprise_server_host`. (Only required if using Github connector for on premises) Host of the Github instance (ex: github.mycompany.com)."
+}
+
+variable "github_enterprise_server_host" {
+  type        = string
+  default     = "" # not null, as this is required if using github_enterprise_server. Validation done on the caller
+  description = "(Only required if using Github Enterprise Server connector) Host of the Github instance (ex: github.mycompany.com)."
 }
 
 variable "github_installation_id" {