+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Changes to be including in future/planned release notes will be added here.
   A Microsoft 365 with admin rights for Entra ID will need to grant admin consent
   for `OnlineMeetings.Read.All` and `OnlineMeetingArtifact.Read.All` permissions.
 
+* GitHub Enterprise Server: variable `github_api_host` is marked as deprecated and will be removed
+  in next major version. Use `github_enterprise_server_host` instead. 
+
+
 ## [0.4.43](https://github.com/Worklytics/psoxy/release/tag/v0.4.43)
  * if you're using the NodeJS test tool, it will be re-installed on your next `terraform apply` due
    to a dependency change.